Server-Sent Events
In order to send messages as a Server Sent Event they need a format
specified in the W3C's spec http://www.w3.org/TR/eventsource/.

Up to now it was left to each developer to format strings according
to this format. With this commit a developer can use a struct which
provides all available fields that a Server Sent Event message
may have.