Add a Plug.Conn.put_resp_body/2 function
This commit adds a `Plug.Conn.put_resp_body/2` function which, well, sets the
response body (`resp_body`) of the connection.

In order to be consistent with `Plug.Conn.resp/3` which sets the state of the
connection to `:set`, this commit also modifies the behavior of `put_status/2`.
Since what a connection needs to be `:set` is a status code and a response body,
now both `put_status/2` and `put_resp_body/2` check for the response body and
the status respectively: if one is set, than setting the "other" one also turns
the state of the connection to `:set`.

This way, these are all equivalents:

    iex> conn |> resp(200, "foo") |> send_resp()
    iex> conn |> put_resp_body("foo") |> put_status(200) |> send_resp()
    iex> conn |> put_status(200) |> put_resp_body("foo") |> send_resp()