Prevent `body_or_params` from overriding content-type
When passing `Plug.Test.conn` a map, the map is passed down to
`body_or_params`, which always assumed the `content-type` would be
`multipart/mixed; charset: utf-8`. If `content-type` is set in the
`conn`'s headers, it will be preserved. If it is not set, it will
default to `multipart/mixex; charset: utf-8`

`body_or_params` now uses `List.keystore` instead of erlang's
`:lists.keystore`