Add :brotli to Plug.Static
Serve [brotli][1] compressed files from Plug.Static. [brotli][1] is an highly
effective compression algorithm, which has a much better compression ratio than
gzip by keeping the decompression speed en par with gzip. It is currently
deployed in Google Chrome and Firefox-44, the WOFF File Format 2.0 and it is
on the way to become an IETF standard.

While decompressing speed is similar to gzip, the same is not true about the
speed of compression: a considerable amount of CPU cycles are added. Thus,
this makes brotli compressed content a perfect candidate to be precompressed
and served via Plug.Static.

[1]: https://github.com/google/brotli