Test.Conn to stop appending chunks after "" sent
Sending a chunk with an empty body kills the connection, however Test.Conn would
still append chunk bodies in test. This commit adds state to track if chunking
is still ongoing, and to stop appending the chunks to the response body if it
has been terminated by sending a chunk with an empty body.