Remove error in Plug.Builder eagerness
Before this change, a Plug.Builder raises an error during compilation when
nothing has been plugged. It makes sense from the standpoint of the
module not performing anything when no plugs have been added to the
pipeline, but raising an error in that case appears to be a little
overly eager.

One must assume that this is done to raise awareness of an unused
Plug.Builder, but that should also be apparent to the programmer given a
lack of calls to plug/1/2. Also there is no way to protect against
plugging a "dummy" function to make the error go away.

This came up in particular while trying to use Plug.Builder in
Phoenix.Endpoint in place of the manual implementation that currently
exists. Endpoints do not ship with any plugs, so compilation fails for
the simplest endpoint uses.