Inflect status from reason phrase
The reason phrase (e.g. Not Found) is now stored against the code in a
map of statuses. These statuses merge those present in both Plug and
Cowboy with the following Exceptions:

In Plug, not in Cowboy:

 * 208 already_reported
 * 308 permanent_redirect
 * 421 misdirected_request
 * 508 loop_detected

In Cowboy, not Plug:

 * 425 - Unordered Collection

Differences (These will cause breaking changes):

 * 226 instance_manipulation_used - "IM Used"
 * 306 - reserved - "Switch Proxy"

The atom to use when sending a response (`put_status(conn, :not_found)`)
is now inflected based on the reason phrase. The code function should
work the same way as before, except it will also work with custom
statuses.

Custom statuses are also displayed in the code helper, however they are
shown after the built in statuses, and will only be displayed if any
have been set.