document custom status codes in docs and error
When reason_phrase is called with an unknown code, an ArgumentError will
be raised with a message describing how to use custom status codes. A
longer version of the documentation in the error message is available in
the moduledoc for conn.ex

In order to prevent compiling status.ex in the tests, some test values
for custom statuses are set in the config when using the test
environment.