Introduce Plug.HTML.html_escape_to_iodata/1
The implementation tries to keep as much of the original string as possible,
splitting only when an escapable character is encountered. This means far fewer
allocations are required and on clean input (that does not need escaping)
means no allocations are performed.
Also the raw iodata is exposed for uses where iodata is accepted anyways,
e.g. in Phoenix.HTML.Safe.

Benchmarks [1] show the new implementation is 1.5 to 9 times faster depending
on input.

escape_char/4 is not inlined and has an empty check for a binary, since
that allows the io_iodata function to be subjected to binary matching
optimisations.

[1]: https://gist.github.com/michalmuskala/1fd844ca83360e18d796131ca41427fa