Add `:set_file` state to Plug.Conn
Using `get_csrf_token/0` and `send_file/3` in the same action is raising
Plug.Conn.AlreadySentError errors. I tracked it down to the before_send
callback that Plug.CSRFProtection registers to put the token into the
session. `send_file/3` was setting the connection status to `:file`,
which was considered "sent" by `put_session/3`.

I added a separate `:set_file` state to be consistent with
`:set_chunked`. I'm unsure of how changing the final state to `:file`
will affect things, so it might be a better option to just add `:file`
to `@unsent`.