Add a prepend_resp_headers/3 to Plug.Conn
At the moment when you want to add another header with an existing key, you would need to manupulate the connection response_headers manually, thus bypassing all internal checks. This function adds a way to the public API of Plug.Conn to add an header that exists already.

this commit fixes #660