Make the adapter contract for push return :ok or {:error, reason}
Since the adapters all use side-effects for push, there is no transformation in
the payload - allowing the original conn to be returned. To maek this explicit,
the contract for `Plug.Adapter` does not allow a transformed payload to be
returned.

To comply with this change, the test adapter has been updated to send a message
on push, and gather the messages when sent_pushes/1 is called.

To ensure that the push messages aren't caught when gathering the response
messages, a 3 tuple of `{ref, :push, asset}` is used.