Refactored the highlighter and added tests
Plug now uses Makeup (if available) to highlight Elixir code in the debugger.
Plug itself does not depend on Makeup and never will, it merely detects
its presence at runtime.

Tests have been updated to reflect the fact that Plug.Debugger now
returns highlighted code (using <span>s with CSS classes), instead of plain text.
Because Makeup is not a dependency, the Debugger must be tested both
with and without the syntax highlighting.

The highlighter has support for manual activation and deactivation,
which can simulate the absence of the Makeup lexers.
That way, although Makeup is installed as a :test dependency,
we can test the debugger as if it weren't.

The fact that we now manually change plug's config in the Debugger's
test module means that it's not completely safe to run in parallel with
other tests.

As a precaution, I've changed the Debugger's test module to use
`async: false`. The flip side is that the test suite now takes ~ 22%
longer to complete.