Replace all Plug.Adapters.Cowboy2 functions with warnings
The Cowboy2 adapter is no longer intended for use. Instead, users should add the
`plug_cowboy` dependency and use that instead.

If the `Plug.Cowboy` module not available then all the functions will show a
warning and then raise. If the module is available then a warning will still be
displayed, encouraging users to use the `Plug.Cowboy` module directly.

The translator test has also been changed to the cowboy 1 specific version as it
is now the responsibility of the plug_cowboy library to add the translator.