Replace all Plug.Adapters.Cowboy functions with warings
Like the Cowboy2 adapter, the Cowboy adapter is no longer intended for use and
required the `plug_cowboy` library. Unlike the Cowboy2 adapter, continued use
will not display a deprecation warning, this is to provide better
maintainability with existing libraries that depend on plug.

The translator and its test have been removed as they are no longer relevant for
plug directly, and should be added by the adapter.