Don't redirect excluded hosts on Plug.SSL (#782)
Previously the exclude list would still redirect http requests to https,
but would not apply the HSTS header. Now the request is no longer
redirected, allowing for requests directly to the excluded hosts to
still work. For example calling `curl http://localhost:4000` when
running in a shell on the server where the application is deployed.