Add actions to exceptions that implement `Plug.Exception`
As to be rendered in the error page.

Adds a `actions/1` function definition to `Plug.Exception` protocol
that should return a list of actions (a map with a `label`, and a `handler`).
Where `label` must be a string and `handler` must be a MFArgs.

Encode and sign the exception actions as to be able to render them in the error page.
Also supports executing an action by matching on a `POST /__plug__/debugger/actions`
that expects the encoded action handler.

Redirects back to the last path after executing the action's handler.
If the last request was not an `GET` it tries to use the `referer` request
header and when all of that fails redirect to `/`.