Improve developer experience within the test adapter (#922)
Related to phoenixframework/phoenix#3605

1. Provide better real-world defaults for custom params in the test adapter
2. Minimize potential compatibility issues

Let's say I have a page containing a list of paginated posts.
I want to write a test that says I can get to page 2 of my
posts, so I write the following in a test:

```elixir
conn = conn(:get, "/posts", page: 2)
```

and then in a handler somewhere, I match on that value:

```elixir
def call(%{params: %{page: page}}, _) when is_integer(page) do
  ...find posts by page...
end
```

At the point, the tests will pass, but they will fail in
production, since the query param values will all be strings.

The current behavior of the test adapter is to treat all
custom params as `body_params`. This has the knock-on effect
of adding a `multipart/mixed` content-type header to the
request when one is not already set.

[RFC 7230 Section 3.3](https://tools.ietf.org/html/rfc7230#section-3.3) states:

> The presence of a message body in a request is signaled by a
Content-Length or Transfer-Encoding header field.  Request message
framing is independent of method semantics, even if the method does
not define any use for a message body.

Rather than require a content-length or transfer-encoding
header to parse custom params as a body, a change which would
likely break a majority of tests in the wild, these changes
aim to provide sensible default behavior for custom params,
regardless of the HTTP method in use.

GET/HEAD request params are _fully stringified_:

```elixir
iex> conn(:get, "/posts", page: 2)
iex> conn.params
%{"page" => "2"}
```

they are **not** fetched as `body_params`:

```elixir
iex> conn.body_params
%Plug.Conn.Unfetched(aspect: :body_params)
```

and finally, no content-type is derived for these requests:

```elixir
iex> conn.req_headers
[]
```

All other behavior remains the same. For instance, binary
values will still be treated as the request body, and
_structs will continue to be passed through unchanged_. This
is to preserve as much backwards compatibility as possible
while still improving developer experience in the simple
case.