add support for rotating cookie session store opts
It's possible to rotate secrets, but also enable or disable encryption,
change key iterations etc.

Writes use main opts so legacy cookies can be migrated transparently.

All options are supported and generally expected.

'secret_key_base' is not copied from the conn struct so it's required in
'rotating_options'.

All options are parsed in 'init', so runtime configuration is limited.