Improve CacheBodyReader example for `:body_reader`
The current example assumes that the body you are reading is going to be
less than 1_000_000 bytes, and would error otherwise.

This version of the code snippet would return the result of
`Plug.Conn.read_body`, so that the `Plug.Parser` can call it again if
needed.

I believe this was the intent of the snippet, as it shows you prepending
the just read body in the assigns. With the current code, I think it
would only ever read the body once.