Plug.Static: Add support for preset MIME types (#1214)
It is not always possible to express custom types with a static
filename -> type mapping. Currently Plug.Static will (if the file
exists) always overwrite the header and halt the connection making it
impossible to preset or later overwrite it based on custom logic.

Thus introduce an opt-out mode to preserve whatever
`Content-Type` was set or not set before.