Customize the Logger metadata key name (#1267)
In certain scenarios we may want to use the Plug.RequestID multiple
times with different configuration options (maybe different headers or
different assigns).

While this is possible, the issue is that every call adds a
`:request_id` key to the Logger metadata. When using the plug muliple
times, the last call will override the metadata of the previous one.

This commit adds a new `:logger_metadata_key` option that allows users to customize
how the request ID will be added to the logger metadata. If not
provided, this option defaults to `:request_id` for backwards
compatibility.