fix: only response with html if client prefers media type
The previous implemtantion has quite basic and didn't account for the q
parameter a client can add to an entry in the accept header. 

This would fail with HTTPie, which send requests with 
"Accept: application/json, text/*;q=0.5". While one can argument that
responding with HTML is in spec, it isn't a good DX.